API-thongtindoanhnghiep/requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
tenacity>=8.2.0
responses>=0.22.0
pytest>=7.0.0